Create composable for fetching posts from endpoint
